Added a comment clarifying the link call in its docstring.
On unix systems, the link() system call uses the convention

 link(const char *oldpath, const char *newpath)

However in java7 (when createLink was added), the createLink call asks
the programmer to use the arguments in the opposite order:

 createLink(newpath, oldpath)

fs/link goes with the java convention, even though the name shadows
the system call's name.  This commit is adding info to the docstring
clarifying which file needs to exist and which one will be created.
It also re-names the arguments to make clear which one must already
exist, and which one is being created.